requirements.txt
absl-py==0.1.11
astroid==1.6.1
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
cliff==2.8.0
cmd2==0.8.0
configparser==3.5.0
cssselect==1.0.3
cycler==0.10.0
fire==0.1.3
html5lib==0.9999999
idna==2.6
isort==4.3.4
kaggle-cli==0.12.13
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.0.0
Markdown==2.6.11
matplotlib==2.2.0
mccabe==0.6.1
MechanicalSoup==0.8.0
numpy==1.14.1
pbr==3.1.1
Pillow==5.0.0
prettytable==0.7.2
progressbar2==3.34.3
protobuf==3.5.2
pycparser==2.18
pylint==1.8.2
pyparsing==2.2.0
pyperclip==1.6.0
python-dateutil==2.7.0
python-utils==2.3.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
stevedore==1.28.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
torch==0.3.1
torchvision==0.2.0
tqdm==4.19.7
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
